+++ CHANGE_LOG.md
@@ -1,5 +1,5 @@
 # v0.3.0
-* Adds fileName, filePath, lineNum, colNum as mandatory tokens for `format()` options
+* Adds fileName, filePath, lineNum, colNum as additional default tokens for `format()` options
 
 # v0.2.0
 * Minor changes